feat: 로그인 형식 확인 코드 간소화 및 적용
